Mouse input scales, new system to get nearby objects in Character

+++ Benefactor/Assets/Scripts/InteractableObject.cs
@@ -95,8 +95,13 @@ public class InteractableObject : MonoBehaviour
 
     public virtual int GetDistance(InteractableObject o)
     {
-        return (int)(Math.Abs(o.transform.position.x - transform.position.x) + Math.Abs(o.transform.position.y - transform.position.y));
+        return (int)Math.Abs(o.transform.position.x - transform.position.x) + (int)Math.Abs(o.transform.position.y - transform.position.y);
     }
+
+    //public virtual int GetDistance(Transform t)
+    //{
+    //    return (int)Math.Abs(t.position.x - transform.position.x) + (int)Math.Abs(t.position.y - transform.position.y);
+    //}
 
     // Update is called once per frame
     void Update()
